backend/requirements.txt
chromadb
fastapi>=0.116.1
google-auth>=2.36.0
itsdangerous>=2.2.0
ollama>=0.4.7
pymupdf>=1.25.3
pydantic-settings>=2.6.1
pdfplumber>=0.11.4
Pillow>=10.4.0
pytesseract>=0.3.13
python-dotenv>=1.0.1
python-multipart>=0.0.20
requests>=2.32.3

numpy==1.26.4

sentence-transformers==3.0.1
transformers==4.41.2

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.0

uvicorn
